// typescript/model/addAnnotationRequest.ts
/**
 * Haven Money API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { BootstrapStrategy } from './bootstrapStrategy';
import { CleanCategory } from './cleanCategory';

export class AddAnnotationRequest {
    'category'?: CleanCategory;
    'memo'?: string;
    'excludeFromBreakdowns'?: boolean;
    'bootstrap': BootstrapStrategy;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "category",
            "baseName": "category",
            "type": "CleanCategory"
        },
        {
            "name": "memo",
            "baseName": "memo",
            "type": "string"
        },
        {
            "name": "excludeFromBreakdowns",
            "baseName": "excludeFromBreakdowns",
            "type": "boolean"
        },
        {
            "name": "bootstrap",
            "baseName": "bootstrap",
            "type": "BootstrapStrategy"
        }    ];

    static getAttributeTypeMap() {
        return AddAnnotationRequest.attributeTypeMap;
    }
}