// typescript/model/balance.ts
/**
 * Haven Money API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class Balance {
    'current': string;
    'available'?: string;
    'limit'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "current",
            "baseName": "current",
            "type": "string"
        },
        {
            "name": "available",
            "baseName": "available",
            "type": "string"
        },
        {
            "name": "limit",
            "baseName": "limit",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return Balance.attributeTypeMap;
    }
}